bugfix: actor, hs and particle systems. further symbol recovery.

+++ src/actor_attach_unit.c
@@ -12,7 +12,7 @@
 #include "headers/blam_data_globals.h"
 
 
-extern void actor_swarm_detach_from_unit(uint16_t actor_index, int unit_index);
+extern void actor_swarm_detach_from_unit(int actor_index, int unit_index);
 extern void actor_delete(int actor_index, uint8_t died);
 extern void actor_detach_from_unit(uint16_t actor_index);
 extern void encounter_attach_unit(int encounter_index, int unit_index);
