+++ src/actor_erase.c
@@ -9,7 +9,7 @@
 #include "headers/blam_data_globals.h"
 
 
-extern void actor_swarm_detach_from_unit(uint16_t actor_index, int unit_index);
+extern void actor_swarm_detach_from_unit(int actor_index, int unit_index);
 extern void object_delete_immediately(int object_index);
 extern void object_delete(int object_index);
 extern void actor_delete(int actor_index, uint8_t died);
