+++ src/unit_died.c
@@ -25,7 +25,7 @@ extern float real_seed_random(uint32_t *seed);
 extern void object_set_garbage(int object_index, uint8_t garbage);
 extern void player_died(int player_index);
 extern void actor_died(int actor_index);
-extern void actor_swarm_unit_died(uint16_t actor_index, int unit_index);
+extern void actor_swarm_unit_died(int actor_index, int unit_index);
 extern int game_time_get(void);
 extern void weapon_owner_update(int weapon_index, uint16_t control_flags, float primary_trigger);
 extern void unit_detach_from_parent(int object_index);
